Migrate Header to Typescript
